+++ src/components/Auth/Lazy.tsx
@@ -1,8 +1,10 @@
-import { h, ComponentType } from 'preact'
+import { h } from 'preact'
 import { localised } from '../../locales'
 import { asyncComponent } from '~utils/components'
 import style from './style.scss'
 import Loader from './assets/loaderSvg'
+
+import type { StepComponentProps } from '~types/routers'
 
 const Loading = () => (
   <div className={style.loading}>
@@ -15,6 +17,6 @@ const AuthCapture = asyncComponent(
   Loading
 )
 
-const AuthLazy = (props: ComponentType<Element>) => <AuthCapture {...props} />
+const AuthLazy = (props: StepComponentProps) => <AuthCapture {...props} />
 
 export default localised(AuthLazy)