+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to the Node [default version scheme](https://docs.npmjs.com
 
 ## [next-version]
 
+## Added
+
+- UI: Added new Intro screen when user begins the Cross Device flow on their mobile device.
+
 ## [6.13.0] - 2021-08-23
 
 ### Added